Create travelling sales men problem.c
The Travelling Salesman Problem (TSP) is a classic problem in computer science that involves finding the shortest possible route that visits a given set of cities and returns to the starting city. It is an NP-hard problem, meaning that it is computationally intractable to find the optimal solution for large problem instances.